% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_cross_marker}
\alias{fname_cross_marker}
\title{Cross marker path}
\usage{
fname_cross_marker(crop = NULL)
}
\arguments{
\item{crop}{string}
}
\value{
character file.path
}
\description{
Cross marker path
}
\author{
Reinhard Simon
}
