% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_dictionary}
\alias{fname_dictionary}
\title{Dictionary path}
\usage{
fname_dictionary(crop = NULL)
}
\arguments{
\item{crop}{string}
}
\value{
character file.path
}
\description{
Dictionary path
}
\author{
Reinhard Simon
}

