% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_material_lists}
\alias{fname_material_lists}
\title{Path to lists of plant materials}
\usage{
fname_material_lists()
}
\value{
character file.path
}
\description{
Path to lists of plant materials
}
\author{
Reinhard Simon
}

