% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_module}
\alias{fname_module}
\title{Module path}
\usage{
fname_module(crop = NULL)
}
\arguments{
\item{crop}{string}
}
\value{
character file.path
}
\description{
Module path
}
\author{
Reinhard Simon
}
