% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_program_stage}
\alias{fname_program_stage}
\title{Breeding program data path}
\usage{
fname_program_stage()
}
\value{
character file.path
}
\description{
Breeding program data path
}
\author{
Reinhard Simon
}

