% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_programs}
\alias{fname_programs}
\title{Breeding program data path}
\usage{
fname_programs()
}
\value{
character file.path
}
\description{
Breeding program data path
}
\author{
Reinhard Simon
}

