% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{get_base_dir}
\alias{get_base_dir}
\title{Base path}
\usage{
get_base_dir(amode = "Default", is_server = FALSE)
}
\arguments{
\item{amode}{mode of data: Local, Demo, BrAPI, ...}

\item{is_server}{logical, default = FALSE}
}
\value{
character file.path or NULL
}
\description{
Base path
}
\author{
Reinhard Simon
}

