% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{fname_fieldbooks}
\alias{fname_fieldbooks}
\title{Fieldbook crops path}
\usage{
fname_fieldbooks()
}
\arguments{
\item{crop}{a crop name}
}
\value{
character file.path
}
\description{
Fieldbook crops path
}
\author{
Reinhard Simon
}

