% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_sites.R
\name{get_base_dir}
\alias{get_base_dir}
\title{Base path}
\usage{
get_base_dir(amode = "Demo")
}
\arguments{
\item{mode}{mode of data: Local, Demo, BrAPI, ...}
}
\value{
character file.path or NULL
}
\description{
Base path
}
\author{
Reinhard Simon
}

